Added function for computing Marcus lift
I added the function `marcus_lift` in `events_rde/paths.py`. This function takes as input a final time `t1`, an array of spike times and a list of spike marks as included in the solution of a call of `SpikingNeuralNetwork`. It will return what is effectively the marcus lift of the path obtained by concatenating the spike train with an extra time dimension. This can then be fed into `signax.signature` to compute the Marcus signature.