Updated `expected_signature_loss`
The previous version of the loss was not very trainable. Since some coordinates of the signature depend on the number of spikes, when comparing spike trains with fundamentally different spike counts, the gradient would end up being in the wrong direction. (The direction in which to step to increase the number of spikes also decreases the spike times of the present spikes)

Now, for two samples of the same size, spike trains are paired up arbitrarily (rather arbitrarily, first sample of the first batch with first sample of the second batch, etc.), and the number of spikes is truncated to the lowest among the pair.