Added support for generating multiple samples at once of `SpikingNeuralNetwork`
Before, if you wanted to generate multiple samples of an instance of `SpikingNeuralNetwork`, you had to wrap the call in `jax.vmap`. It turns out that this is very slow, espicially when computing gradients. This is probably because the computation graph is different for each sample and so there is no significant speed up ofter compilation.

In the new approach, to generate multiple sample, each block of solvers between event times (i.e., inside the body function of the while loop) is wrapped in a `jax.vmap` and the while loop now continous until all of them have reached `t1`. Even though this means that we call `diffrax.diffeqsolve` on some samples that have already reached `t1` (we have to wait for the last one, and each sample has a different number of spikes), this is still *a lot* faster than the previous approach.